+++ lib/vagrant-omnibus/action.rb
@@ -20,7 +20,7 @@ module VagrantPlugins
   module Omnibus
     module Action
       autoload :InstallChef, File.expand_path("../action/install_chef", __FILE__)
-      autoload :IsRunning, File.expand_path("../action/is_running", __FILE__)
+      autoload :IsRunningOrActive, File.expand_path("../action/is_running_or_active", __FILE__)
       autoload :ReadChefVersion, File.expand_path("../action/read_chef_version", __FILE__)
 
       # Include the built-in modules so that we can use them as top-level
@@ -31,7 +31,7 @@ module VagrantPlugins
       def self.install_chef
         @install_chef ||= ::Vagrant::Action::Builder.new.tap do |b|
           b.use ConfigValidate
-          b.use Call, IsRunning do |env1, b2|
+          b.use Call, IsRunningOrActive do |env1, b2|
             if env1[:result]
               b2.use ReadChefVersion
               b2.use InstallChef