+++ CHANGELOG.md
@@ -13,6 +13,7 @@ BUG FIXES:
   string. fixes [GH-2] for realz.
 * RubyGems 2.0 compat: use #empty? to check for results. fixes [GH-6]
 * Ensure 'vagrant-rackspace/action' is loaded. fixes [GH-7]
+* Trigger plugin if machine state is `:active`. fixes [GH-8]
 
 ## 1.0.1 (April 17, 2013)